+++ internal/web/templates/pages/admin_post_form.tmpl
@@ -1,6 +1,10 @@
 {{ define "pages/admin_post_form.tmpl" }}
   {{ template "partials/header" . }}
   <h1>{{ .Title }}</h1>
+  <p>
+    <a href="/admin/posts">返回文章列表</a>
+    · <a href="/admin">返回后台</a>
+  </p>
   <form method="post" action="{{ .Action }}">
     <input type="hidden" name="_csrf" value="{{ .CSRFToken }}">
     <div>
