+++ internal/web/templates/partials/header.tmpl
@@ -19,6 +19,9 @@
     <nav class="nav">
       <a href="/">首页</a>
       <a href="/archive">归档</a>
+      <a href="/categories">分类</a>
+      <a href="/tags">标签</a>
+      <a href="/admin">管理</a>
     </nav>
   </header>
   <main class="container">
